Leads AI Agent/requirements.txt
Flask==2.2.2
flask-cors==3.0.10
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0
Faker==18.9.0
transformers==4.30.0
torch==2.0.1
chromadb==0.3.21
langchain==0.0.142
streamlit==1.25.0
python-dotenv==1.0.0
